backend/requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
Flask-CORS>=4.0.0
google-api-python-client>=2.111.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
Flask-APScheduler>=1.13.1
pandas>=2.1.3
Flask-JWT-Extended>=4.5.3
scikit-learn>=1.3.2
fpdf2>=2.7.6
matplotlib>=3.8.2
xlsxwriter>=3.1.9
python-dotenv>=1.0.1
gunicorn>=21.2.0